Match-id-89976f9999f3025cf6acc4fffccbd69a01aeeb70

+++ src/AccCTR/src/unique/unique_func.cpp
@@ -119,10 +119,12 @@ void Dedup::NewParameter()
         // Time to check the proper size of sharded tables for performance
         // sake.
         uint64_t shardedTableSize = 0;
-        if (std::numeric_limits<uint64_t>::max() / n / groupCount_ < newBucketCountPowerOf2) {
-            shardedTableSize = std::numeric_limits<int>::max();
+        if (std::numeric_limits<uint64_t>::max() / static_cast<uint64_t>(n) /
+        static_cast<uint64_t>(groupCount_) <
+        newBucketCountPowerOf2) {
+            shardedTableSize = static_cast<uint64_t>(std::numeric_limits<int>::max());
         } else {
-            shardedTableSize = newBucketCountPowerOf2 * n * groupCount_;
+            shardedTableSize = newBucketCountPowerOf2 * n * static_cast<uint64_t>(groupCount_);
         }
 
         int largeCount = 0;
